+++ breeze-core/src/main/kotlin/extensions/TextExtensions.kt
@@ -5,129 +5,137 @@
 
 package com.windea.breezeframework.core.extensions
 
-//这些特殊的字符串处理方法相比let & joinToString更加灵活，更加适用于一般情况
-//允许更多自定义项
-
 /**
- * 使用特定的转化规则，将当前对象转化成字符串。
- * * 空值 -> 返回空字符串。
- * * 存在转化方法 -> 使用该方法将当前对象转化为字符串。
- * * 其他情况 -> 使用默认方法将当前对象转化为字符串。
+ * 将当前对象转化成文本。
+ * * 如果当前对象是空值，则返回空字符串。
+ * * 如果存在转化方法，则使用该方法将当前对象转化为字符串。
  */
-fun <T : Any> T?.typing(transform:((T) -> String)? = null):String {
+fun <T : Any> T?.toText(transform:((T) -> String)? = null):String {
 	return if(this == null) "" else if(transform != null) transform(this) else toString()
 }
 
 
 /**
- * 使用特定的转化规则，将当前数组转化成字符串。
- * * 忽略空元素（默认为`true`） && 存在空元素 -> 忽略该元素.
- * * 忽略空数组（默认为`true`） && 不存在有效的元素 -> 返回空字符串。
- * * 存在转化方法 -> 使用该方法将元素转化为字符串。
- * * 其他情况 -> 使用默认方法将元素转化为字符串。
+ * 将当前数组拼接并转化成文本。
+ * 默认忽略空数组，忽略空元素，没有缩进。
+ * * 允许指定前缀、后缀、分隔符和缩进。
+ * * 如果忽略空数组，且元素全部为空值，则返回空字符串。
+ * * 如果忽略空元素，且元素为空值，则忽略该元素。
+ * * 如果存在转化方法，则使用该方法将元素转化为字符串。
  */
-fun <T : Any> Array<out T?>.typingAll(
-	separator:CharSequence = ", ", prefix:CharSequence = "", postfix:CharSequence = "", indent:String? = null,
-	omitEmptyElement:Boolean = true, omitEmpty:Boolean = true,
-	transform:((T) -> CharSequence)? = null
+fun <T : Any> Array<out T?>.joinToText(
+	separator: CharSequence = ", ",
+	prefix: CharSequence = "",
+	postfix: CharSequence = "",
+	indent: CharSequence = "",
+	omitEmpty: Boolean = true,
+	omitEmptyElement: Boolean = true,
+	transform: ((T) -> CharSequence)? = null
 ):String {
 	var result = buildString {
 		var count = 0
-		for(element in this@typingAll) {
+		for(element in this@joinToText) {
 			val snippet = if(element == null) null else if(transform != null) transform(element) else element.toString()
 			if(omitEmptyElement && snippet.isNullOrEmpty()) continue
 			if(count++ > 0) append(separator)
 			append(snippet)
 		}
 	}
-	if(indent != null) result = result.prependIndent(indent)
+	if(indent.isNotEmpty()) result = result.prependIndent(indent.toString())
 	if(omitEmpty && result.isNotEmpty()) result = "$prefix$result$postfix"
 	return result
 }
 
 /**
- * 使用特定的转化规则，将当前集合转化成字符串。
- * * 忽略空元素（默认为`true`） && 存在空元素 -> 忽略该元素.
- * * 忽略空集合（默认为`true`） && 不存在有效的元素 -> 返回空字符串。
- * * 存在转化方法 -> 使用该方法将元素转化为字符串。
- * * 其他情况 -> 使用默认方法将元素转化为字符串。
+ * 将当前集合拼接并转化成文本。
+ * 默认忽略空集合，忽略空元素，没有缩进。
+ * * 允许指定前缀、后缀、分隔符和缩进。
+ * * 如果忽略空集合，且元素全部为空值，则返回空字符串。
+ * * 如果忽略空元素，且元素为空值，则忽略该元素。
+ * * 如果存在转化方法，则使用该方法将元素转化为字符串。
  */
-fun <T : Any> Iterable<T?>.typingAll(
-	separator:CharSequence = ", ", prefix:CharSequence = "", postfix:CharSequence = "", indent:String? = null,
-	omitEmpty:Boolean = true, omitEmptyElement:Boolean = true,
-	transform:((T) -> CharSequence)? = null
+fun <T : Any> Iterable<T?>.joinToText(
+	separator: CharSequence = ", ",
+	prefix: CharSequence = "",
+	postfix: CharSequence = "",
+	indent: CharSequence = "",
+	omitEmpty: Boolean = true,
+	omitEmptyElement: Boolean = true,
+	transform: ((T) -> CharSequence)? = null,
 ):String {
 	var result = buildString {
 		var count = 0
-		for(element in this@typingAll) {
+		for(element in this@joinToText) {
 			val snippet = if(element == null) null else if(transform != null) transform(element) else element.toString()
 			if(omitEmptyElement && snippet.isNullOrEmpty()) continue
 			if(count++ > 0) append(separator)
 			append(snippet)
 		}
 	}
-	if(indent != null) result = result.prependIndent(indent)
+	if(indent.isNotEmpty()) result = result.prependIndent(indent.toString())
 	if(omitEmpty && result.isNotEmpty()) result = "$prefix$result$postfix"
 	return result
 }
 
 /**
- * 使用特定的转化规则，将当前映射转化成字符串。
- * * 忽略空键值对（默认为`true`） && 存在空键值对 -> 忽略该键值对.
- * * 忽略空值（默认为`true`） && 存在空值 -> 忽略该键值对.
- * * 忽略空映射（默认为`true`） && 不存在有效的键值对或值 -> 返回空字符串。
- * * 存在转化方法 -> 使用该方法将元素转化为字符串。
- * * 其他情况 -> 使用默认方法将元素转化为字符串。
+ * 将当前映射拼接并转化成文本。
+ * 默认忽略空映射，忽略空值，没有缩进。
+ * * 允许指定前缀、后缀、分隔符和缩进。
+ * * 如果忽略空映射，且键值对的值全部为空值，则返回空字符串。
+ * * 如果忽略空值，且键值对的值为空值，则忽略该键值对。
+ * * 如果存在转化方法，则使用该方法将键值对转化为字符串。
  */
-fun <K, V> Map<K, V>.typingAll(
-	separator:CharSequence = ", ", prefix:CharSequence = "", postfix:CharSequence = "", indent:String? = null,
-	omitEmpty:Boolean = true, omitEmptyEntry:Boolean = true, omitEmptyValue:Boolean = true,
-	transform:((Map.Entry<K, V>) -> CharSequence)? = null
+fun <K, V> Map<K, V>.joinToText(
+	separator: CharSequence = ", ",
+	prefix: CharSequence = "",
+	postfix: CharSequence = "",
+	indent: CharSequence = "",
+	omitEmpty: Boolean = true,
+	omitEmptyValue: Boolean = true,
+	transform: ((Map.Entry<K, V>) -> CharSequence)? = null,
 ):String {
 	var result =  buildString {
 		var count = 0
-		for(entry in this@typingAll) {
+		for(entry in this@joinToText) {
 			val valueSnippet = entry.value?.toString()
 			if(omitEmptyValue && valueSnippet.isNullOrEmpty()) continue
 			val snippet = if(transform == null) entry.toString() else transform(entry)
-			if(omitEmptyEntry && snippet.isEmpty()) continue
 			if(count++ > 0) append(separator)
 			append(snippet)
 		}
-		if(omitEmpty && length > 0) insert(0, prefix).append(postfix)
 	}
-	if(indent != null) result = result.prependIndent(indent)
+	if(indent.isNotEmpty()) result = result.prependIndent(indent.toString())
 	if(omitEmpty && result.isNotEmpty()) result = "$prefix$result$postfix"
 	return result
 }
 
 /**
- * 使用特定的转化规则，将当前映射转化成字符串。
- * * 没有有效值的序列 && 忽略空虚列  -> 返回空字符串
- * * 空值和空字符串 && 忽略空元素 -> 忽略该元素
- * * 存在转化方法 -> 使用转化方法转化元素
+ * 将当前序列拼接并转化成文本。
+ * 默认忽略空序列，忽略空元素，没有缩进。
+ * * 允许指定前缀、后缀、分隔符和缩进。
+ * * 如果忽略空序列，且元素全部为空值，则返回空字符串。
+ * * 如果忽略空元素，且元素为空值，则忽略该元素。
+ * * 如果存在转化方法，则使用该方法将元素转化为字符串。
  */
-fun <T : Any> Sequence<T?>.typingAll(
-	separator:CharSequence = ", ", prefix:CharSequence = "", postfix:CharSequence = "", indent:String? = null,
-	omitEmpty:Boolean = true, omitEmptyElement:Boolean = true,
-	transform:((T) -> CharSequence)? = null
+fun <T : Any> Sequence<T?>.joinToText(
+	separator: CharSequence = ", ",
+	prefix: CharSequence = "",
+	postfix: CharSequence = "",
+	indent: CharSequence = "",
+	omitEmpty: Boolean = true,
+	omitEmptyElement: Boolean = true,
+	transform: ((T) -> CharSequence)? = null,
 ):String {
 	var result =  buildString {
 		var count = 0
-		for(element in this@typingAll) {
+		for(element in this@joinToText) {
 			val snippet = if(element == null) null else if(transform != null) transform(element) else element.toString()
 			if(omitEmptyElement && snippet.isNullOrEmpty()) continue
 			if(count++ > 0) append(separator)
 			append(snippet)
 		}
-		if(omitEmpty && length > 0) insert(0, prefix).append(postfix)
 	}
-	if(indent != null) result = result.prependIndent(indent)
+	if(indent.isNotEmpty()) result = result.prependIndent(indent.toString())
 	if(omitEmpty && result.isNotEmpty()) result = "$prefix$result$postfix"
 	return result
 }
-
-
-//fun String.typingToList(separator:CharSequence= ", ",prefix:CharSequence = "",postfix:CharSequence=""):List<String> = TODO()
-
-//fun String.typingToSequence(separator:CharSequence = ", ",prefix:CharSequence = "",postfix:CharSequence=""):Sequence<String> = TODO()